backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
aiofiles==24.1.0
requests==2.32.3
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv